requirements.txt
discord.py~=2.6.4
pandas~=3.0.1
PyMySQL~=1.1.2
SQLAlchemy~=2.0.46
